More on JPA integration. Login working More fixes to come. SessionClientFromLocal still broken.

+++ unm-backend-domain/src/main/java/fr/univmobile/backend/domain/CategoryRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface CategoryRepository extends CrudRepository<Category, Integer> {
+public interface CategoryRepository extends CrudRepository<Category, Long> {
 	
 	List<Category> findByName(String name);
