+++ unm-backend-domain/src/main/java/fr/univmobile/backend/domain/ImageMapRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ImageMapRepository extends CrudRepository<ImageMap, Integer> {
+public interface ImageMapRepository extends CrudRepository<ImageMap, Long> {
 	
 	List<ImageMap> findByName(String name);
